Speculative fix for perf of having 10 simultaneous cards.
Don't rendcer the messages in cards not at the top and don't render the older messages in a thread.
We still fetch all the messages though, so it should still
benefit faster swiping through cards since there will be less time blocked on the network fetching messages.
Although, that could slow down initial triage actions, so could be more nuanced.